requirements.txt
# Core Web Framework
fastapi==0.95.1
uvicorn==0.22.0

# Database
chromadb==0.4.3
sqlalchemy==2.0.19
alembic==1.11.1

# Machine Learning and NLP
llama-cpp-python==0.2.20
openai-whisper==20231117

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and Utilities
typing-extensions==4.6.3

# Development and Testing
pytest==7.3.1
httpx==0.24.1

# Optional: For better performance
numpy==1.24.3

# Video processing
yt-dlp==2023.7.6

# Keyword extraction and NLP
spacy==3.6.1
scikit-learn==1.3.0